# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HostYourHackathon API",
        "description": "REST API documentation for HostYourHackathon platform",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/hackathons/{hackathon}/submissions": {
            "post": {
                "tags": [
                    "Submissions"
                ],
                "summary": "Create a submission (Leader only)",
                "operationId": "createSubmission",
                "parameters": [
                    {
                        "name": "hackathon",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "problem_statement",
                                    "description"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "problem_statement": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "tech_stack": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "demo_url": {
                                        "type": "string",
                                        "format": "url",
                                        "nullable": true
                                    },
                                    "repo_url": {
                                        "type": "string",
                                        "format": "url",
                                        "nullable": true
                                    },
                                    "is_draft": {
                                        "type": "boolean",
                                        "default": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation or business logic failed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/submissions/{submission}": {
            "get": {
                "tags": [
                    "Submissions"
                ],
                "summary": "Get submission details",
                "operationId": "getSubmission",
                "parameters": [
                    {
                        "name": "submission",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Submissions"
                ],
                "summary": "Update a draft submission (Leader only)",
                "operationId": "updateSubmission",
                "parameters": [
                    {
                        "name": "submission",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "problem_statement",
                                    "description"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "problem_statement": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "tech_stack": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "demo_url": {
                                        "type": "string",
                                        "format": "url",
                                        "nullable": true
                                    },
                                    "repo_url": {
                                        "type": "string",
                                        "format": "url",
                                        "nullable": true
                                    },
                                    "is_draft": {
                                        "type": "boolean",
                                        "default": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/submissions/{submission}/finalize": {
            "post": {
                "tags": [
                    "Submissions"
                ],
                "summary": "Finalize a submission (Leader only)",
                "operationId": "finalizeSubmission",
                "parameters": [
                    {
                        "name": "submission",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "422": {
                        "description": "Failed to finalize"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/submissions/{submission}/files": {
            "post": {
                "tags": [
                    "Submissions"
                ],
                "summary": "Upload a file to a submission",
                "operationId": "uploadSubmissionFile",
                "parameters": [
                    {
                        "name": "submission",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/submissions/{submission}/files/{file}": {
            "delete": {
                "tags": [
                    "Submissions"
                ],
                "summary": "Delete a submission file",
                "operationId": "deleteSubmissionFile",
                "parameters": [
                    {
                        "name": "submission",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "file",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "422": {
                        "description": "Failed to delete"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}